Added main inputs, view port settings, and basic movement
Added inputs to the project settings for the basic character actions, along with an input manager node and script to track, store, and pass variables to the things that need it (mainly the player). added a script to the player that reads the input manager and uses those variables to move. Tweaked the window settings so that the visuals are scaled to x2 size, and the window size is about double what I'd put in a browser by default. It seems to play well in it's default size as well as full-screen.