fixed player collision, set up enemy behavior script
Fixed player collision in it's movement function to utilize velocity and move_and_slide. Re-worked both the player and enemy scene hierarchy so that the base node is the correct type. Set up the basic hierarchy for the skeleton enemy, including the character animator script, and the variables needed for that script to work within the enemy behavior script.