Improved movement and damage behaviors, added knockback and camera follow
changed hazard area script to check for a variable before applying damage so that it doesn't attempt to damage non-player characters, and moved the damage call into process so that it will apply damage as soon as the player's invulnerability ends if they are still within the area. added some knock-back mechanics to damage, and moved movement and knockback logic into their own functions. added position anchor nodes to the hazard and player to get more accurate calculations for knockback direction. Added camera and camera behavior to game manager so that it lerps to the player's global position, but only if we want it to (and we will most of the time). This should let us lerp somewhere else if we want to shift our view. added sprite flickering behavior to sprite animator script and player.