Rearranged scene files, added check point system
Rearranged some of the scene files into the proper folder. Added more robust behavior for the game manager, including a check point manager, world, and region scripts. Added check point scene and scripts, and a system for registering check points based on the region scene they were in when the region was loaded (and their index). For now, there isn't any behavior in the respawn system for loading the scene before transporting to the check-point, but that's next in the pipe-line. Basic respawn behavior set up which triggers automatically after death, healing the player and warping them to a check point's anchor.