Added transporter, added level loading, expanded checkpoint system
Centralized more of the game's initial setup through the game manager and world manager so that fewer things are being independently called through ready that may be order-sensitive. Added a transporter object that in-conjunction with the region node, allows the player to be transported to a specific scene's transporter (by scene index and transporter index). Also expanded the check-point system to add scene loading and unloading to allow the player to respawn at a check point from another region (it was initially built to do that anyways). Haven't added transition effects, so the camera movement is currently pretty jarring, but it's on the to-do list.