skeleton behavior complete-ish (chase, gloat, return)
added player detection area to the enemy, though it doesn't rotate to match the facing yet since animation facing isn't a thing for the skeleton yet. Added other enemy state behaviors to the move method including gloating, chase, and return. The skeleton will gloat after hitting the player, and after gloating they will either continue chasing or return depending on if you are outside of the pursuit range relative to the patrol zone. The skeleton will also stop chasing as soon as the player is outside of the pursuit range and switch to return. May set it up so that the skeleton switches to pursuit if the player enters the patrol zone while the skeleton is in return mode.