enemy detection zone rotation, enemy movement, cancel return behavior
Enemy detection area now correctly rotates based on the snapped direction of the most recent enemy movement call. enemy now animates based on whether or not they are currently moving, and depending on what direction their most recent movement call was in, playing idle animations during gloat and detect timers and switching to walking animations while moving. The enemy will switch from return to detect if the player manages to get within the patrol area (closer to the center of all of the patrol points than the closest point is to the center) while the enemy is returning, basically resulting in an automatic notice if the enemy's return speed gets reduced enough to make duping the enemy possible.