added conductor, performer, and timed flame jet
Added y sorting to the relevant scenes to allow skeletons and player to y sort (it has to be enabled on the scene's parent node).
Added a conductor script (maestro) and a performer script for timing actions to musical beats.
Added flame hazard object, it's animation trees, and connected it to performer.