ice and rough tile behaviors
added ice area script that can be applied to any area 2d, including polygon area (perfect for overlaying a map layer). added a rough island script that overrides the slipping bool (for small islands) so that I don't have to fiddle with ice colliders as much for islands of non-ice.