started level design phase
Added collision physics layer to main tile map, and saved it as it's own tile map (to re-use in other rooms). Created a basic level 1 tutorial area, and re-set it as the first loaded region instead of the debug levels. Created the start of a level 2 area, added it to the world manager, and linked the zones via stairs. created a script to display text on area enter (for entrance stairs, to tell the player that they don't want to leave the dungeon). Created different versions of the stairs scene for the wider stair cases.
Moved push pot into scenes.